requirements.txt
# requirements.txt
Django==1.8.2
djangorestframework==3.1.2
djangorestframework-jwt==1.5.0
PyJWT==1.3.0
Markdown==2.6.2
stripe==1.22.2
django-jsonfield==0.9.10
django-suit==0.2.3
django-debug-toolbar==0.9.4
gunicorn
django-codemirror-widget==0.4.0